Extract reader and writer classes from MoneySupport
